post, category api 제거, static 서비스 구현
