--- env_props
@@ -1,1 +1,0 @@
-NEXT_PUBLIC_API_PATH=
